+++ webforum/app/Http/Controllers/ObjavaController.php
@@ -10,6 +10,8 @@ use Illuminate\Support\Facades\Validator;
 
 use Carbon\Carbon;
 
+use Illuminate\Support\Facades\Auth;
+
 class ObjavaController extends Controller
 {
     public function index()
@@ -28,18 +30,29 @@ class ObjavaController extends Controller
     public function store(Request $request)
     {
 
+        $user_id = Auth::user()->id; 
+
     $validator = Validator::make($request->all(), [
         'naziv' => 'required',
         'tekst' => 'required',
         'tema_id' => 'required',
-        'user_id' => 'required',
+
 
     ]);
 
     if ($validator->fails()) {
         return response()->json($validator->errors());
     }
-
+            //MODERATOR TEMA
+            $jeModeratorTeme = Auth::user()->jeModeratorTeme;
+            //MODERATOR ZAJEDNICA
+            $jeModeratorZajednice = Auth::user()->jeModeratorZajednice;
+            //ADMINISTRATOR
+            $jeAdmin = Auth::user()->jeAdmin;
+    
+            if ($jeModeratorTeme || $jeModeratorZajednice || $jeAdmin) {
+                return response()->json(['error' => 'NEOVLASCEN PRISTUP: Administrator i moderatori nemaju ovlascenje da kreiraju objave'], 403);
+            }
 
     $objava = new Objava();
     $objava->naziv = $request->naziv;
@@ -48,7 +61,7 @@ class ObjavaController extends Controller
     $objava->brojSvidjanja = 0;
     $objava->brojNesvidjanja = 0;
     $objava->tema_id = $request->tema_id;
-    $objava->user_id = $request->user_id;
+    $objava->user_id = $user_id;
 
     $objava->save();
 
@@ -59,6 +72,14 @@ class ObjavaController extends Controller
 
     public function updateTekst(Request $request, $id)
      {
+        $user_id = Auth::user()->id; 
+        $jeAdmin = Auth::user()->jeAdmin;
+        $objava_user_id = Objava::where('id', $id)->value('user_id');
+
+        if($user_id != $objava_user_id || !$jeAdmin){
+            return response()->json(['error' => 'NEOVLASCEN PRISTUP: Objavu mogu menjati ili korisnik koji ju je kreirao ili admin!'], 403);
+        }
+
          $request->validate([
              'tekst' => 'required'
          ]);
@@ -74,6 +95,16 @@ class ObjavaController extends Controller
 
     public function destroy($id)
     {
+
+        $user_id = Auth::user()->id; 
+        $jeAdmin = Auth::user()->jeAdmin;
+        $objava_user_id = Objava::where('id', $id)->value('user_id');
+
+        if($user_id != $objava_user_id || !$jeAdmin){
+            return response()->json(['error' => 'NEOVLASCEN PRISTUP: Objavu mogu brisati ili korisnik koji ju je kreirao ili admin!'], 403);
+        }
+
+
         $objava = objava::findOrFail($id);
         $objava->delete();
         return response()->json('Data objava je uspesno obrisana!');
