+++ webforum/app/Http/Controllers/ZajednicaController.php
@@ -8,6 +8,8 @@ use App\Http\Resources\ZajednicaResource;
 use App\Models\Zajednica;
 use Illuminate\Support\Facades\Validator;
 
+use Illuminate\Support\Facades\Auth;
+
 class ZajednicaController extends Controller
 {
     public function index()
@@ -26,10 +28,11 @@ class ZajednicaController extends Controller
     public function store(Request $request)
     {
 
+        $user_id = Auth::user()->id; 
+
     $validator = Validator::make($request->all(), [
         'naziv' => 'required',
         'opis' => 'required',
-        'user_id' => 'required',
 
     ]);
 
@@ -37,12 +40,24 @@ class ZajednicaController extends Controller
         return response()->json($validator->errors());
     }
 
+    
+                //MODERATOR TEMA
+                $jeModeratorTeme = Auth::user()->jeModeratorTeme;
+                //MODERATOR ZAJEDNICA
+                $jeModeratorZajednice = Auth::user()->jeModeratorZajednice;
+                //ADMINISTRATOR
+                $jeAdmin = Auth::user()->jeAdmin;
+        
+                if ($jeModeratorTeme || $jeModeratorZajednice || $jeAdmin) {
+                    return response()->json(['error' => 'NEOVLASCEN PRISTUP: Administrator i moderatori nemaju ovlascenje da kreiraju zajednice'], 403);
+                }
+
 
     $zajednica = new Zajednica();
     $zajednica->naziv = $request->naziv;
     $zajednica->opis = $request->opis;
     $zajednica->brojTema = 0;
-    $zajednica->user_id = $request->user_id;
+    $zajednica->user_id = $user_id;
 
     $zajednica->save();
 
@@ -52,10 +67,18 @@ class ZajednicaController extends Controller
 
     public function update(Request $request, $id)
     {
+
+        $user_id = Auth::user()->id; 
+        $jeModeratorZajednice = Auth::user()->jeModeratorZajednice;
+        $zajednica_user_id = Zajednica::where('id', $id)->value('user_id');
+
+        if($user_id != $zajednica_user_id || !$jeModeratorZajednice){
+            return response()->json(['error' => 'NEOVLASCEN PRISTUP: Zajednicu mogu menjati samo moderator zajednica ili korisnik koji ju je kreirao!'], 403);
+        }
+
         $validator = Validator::make($request->all(), [
             'naziv' => 'required',
             'opis' => 'required',
-            'user_id' => 'required',
         ]);
 
         if ($validator->fails()) {
@@ -68,7 +91,7 @@ class ZajednicaController extends Controller
         $zajednica->naziv = $request->naziv;
         $zajednica->opis = $request->opis;
         $zajednica->brojTema = $request->brojtema;
-        $zajednica->user_id = $request->user_id;
+        $zajednica->user_id = $user_id;
 
         $zajednica->save();
 
@@ -78,6 +101,15 @@ class ZajednicaController extends Controller
 
     public function updateOpis(Request $request, $id)
      {
+
+        $user_id = Auth::user()->id; 
+        $jeModeratorZajednice = Auth::user()->jeModeratorZajednice;
+        $zajednica_user_id = Zajednica::where('id', $id)->value('user_id');
+
+        if($user_id != $zajednica_user_id || !$jeModeratorZajednice){
+            return response()->json(['error' => 'NEOVLASCEN PRISTUP: Zajednicu mogu menjati samo moderator zajednica ili korisnik koji ju je kreirao!'], 403);
+        }
+
          $request->validate([
              'opis' => 'required'
          ]);
@@ -93,6 +125,15 @@ class ZajednicaController extends Controller
 
     public function destroy($id)
     {
+
+        $user_id = Auth::user()->id; 
+        $jeModeratorZajednice = Auth::user()->jeModeratorZajednice;
+        $zajednica_user_id = Zajednica::where('id', $id)->value('user_id');
+
+        if($user_id != $zajednica_user_id || !$jeModeratorZajednice){
+            return response()->json(['error' => 'NEOVLASCEN PRISTUP: Zajednicu mogu brisati samo moderator zajednica ili korisnik koji ju je kreirao!'], 403);
+        }
+
         $zajednica = zajednica::findOrFail($id);
         $zajednica->delete();
         return response()->json('Data zajednica je uspesno obrisana!');
